Conflicts:
	pf/WebContent/WEB-INF/editarPersona.jsp

+++ pf/src/servlet/admin/ListarPersonas.java
@@ -2,6 +2,7 @@ package servlet.admin;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.LinkedList;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -14,6 +15,7 @@ import datos.PersonasDAO;
 import entidad.ApplicationException;
 import entidad.Institucion;
 import entidad.Persona;
+import entidad.TipoPersona;
 
 /**
  * Servlet implementation class ListarPersonas
@@ -61,12 +63,32 @@ public class ListarPersonas extends HttpServlet {
 						+ "/reguistarPersonas");
 
 			} else if (request.getParameter("eliminar") != null) {
-
+				
 				int id = Integer.parseInt(request.getParameter("eliminar"));
 				persona = perDao.buscarPersonaId(id);
-				perDao.eliminarPersona(persona);
-				response.sendRedirect(request.getContextPath()
-						+ "/admin/listarPersonas");
+			    if(persona.getTipoPersona().getIdTipoPersona()== TipoPersona.ADMINISTADOR){
+			    	LinkedList<Persona> administrador =perDao.buscarPersonaTipoPersona(TipoPersona.ADMINISTADOR);
+			    	if(administrador!=null && administrador.size()>2){
+			    		perDao.eliminarPersona(persona);
+						response.sendRedirect(request.getContextPath()
+								+ "/admin/listarPersonas");
+			    	}else {
+						request.getSession().setAttribute("error", "Debe existir al menos un administrador");
+						response.sendRedirect(request.getContextPath()
+								+ "/admin/listarPersonas");
+					}
+			    	
+			    	
+			    }else {
+			    	perDao.eliminarPersona(persona);
+					response.sendRedirect(request.getContextPath()
+							+ "/admin/listarPersonas");
+			    }
+				
+
+				
+				
+				
 
 			} else if (request.getParameter("eliminar") == null && request.getParameter("editar") == null  ){
 				response.sendRedirect(request.getContextPath()
@@ -75,20 +97,24 @@ public class ListarPersonas extends HttpServlet {
 
 		} catch (IOException | NumberFormatException | SQLException ex) {
 			// TODO: handle exception
+			
+			request.getSession().setAttribute("error", "Ocurrio un error inesperado");
 			response.sendRedirect(request.getContextPath()
 					+ "/admin/listarPersonas");
 		}catch (ApplicationException ex) {
 			// TODO: handle exception
+		
 			request.getSession().setAttribute("error", ex.getMessage());
 			response.sendRedirect(request.getContextPath()
 					+ "/admin/listarPersonas");
 		}catch (Exception ex) {
 			// TODO: handle exception
+			
 			request.getSession().setAttribute("error", "Ocurrio un error inesperado");
 			response.sendRedirect(request.getContextPath()
 					+ "/admin/listarPersonas");
 		}
-
+		System.out.println(request.getSession().getAttribute("error"));
 	}
 
 }
